refactor/AOS: 패키지명 변경 - ComponentUtil to component
